Move top level commands to own subpackages
