+++ src/snowcli/cli/warehouse/commands.py
@@ -1,12 +1,11 @@
 from __future__ import annotations
 
 import typer
-from snowflake.connector.cursor import SnowflakeCursor
 
 from snowcli.cli.common.decorators import global_options_with_connection
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
+from snowcli.cli.warehouse.manager import WarehouseManager
 from snowcli.output.decorators import with_output
-from snowcli.cli.common.sql_execution import SqlExecutionMixin
 from snowcli.output.printing import OutputData
 
 app = typer.Typer(
@@ -14,11 +13,6 @@ app = typer.Typer(
     context_settings=DEFAULT_CONTEXT_SETTINGS,
     help="Manage warehouses",
 )
-
-
-class WarehouseManager(SqlExecutionMixin):
-    def show(self) -> SnowflakeCursor:
-        return self._execute_query("show warehouses")
 
 
 @app.command("status")
